src/frauddetection/requirements.txt
flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
google-cloud-core==2.3.3
PyJWT==2.8.0
cryptography==41.0.7
numpy==1.24.4
pandas==2.0.3
structlog==23.2.0
werkzeug==2.3.7
pytest==7.4.3
pytest-flask==1.3.0
mock==5.1.0
autopep8==2.0.4
flake8==6.1.0
